WorldMap::GetArea: use size_t for indexing

+++ gemrb/core/GUI/WorldMapControl.cpp
@@ -177,8 +177,8 @@ void WorldMapControl::ScrollTo(const Point& pos)
 
 	if (pos.IsZero()) {
 		// center worldmap on current area
-		unsigned entry;
-		const WMPAreaEntry *areaEntry = worldmap->GetArea(currentArea, entry);
+		size_t entry;
+		const WMPAreaEntry* areaEntry = worldmap->GetArea(currentArea, entry);
 		if (areaEntry) {
 			Pos.x = areaEntry->pos.x - frame.w / 2;
 			Pos.y = areaEntry->pos.y - frame.h / 2;
